pyproject.toml
[tool.poetry]
name = "projects"
version = "0.1.0"
repository = "https://github.com/lindseygulden/leg-up"
description = "Data science projects"
authors = ["Lindsey Gulden <lindsey.e.gulden@gmail.com>"]
license = "GNU 3.0"
readme = "README.md"
packages = [{ include = "projects" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pre-commit = "^3.3.3"
pytest = "^7.4.0"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
scikit-learn = "^1.3.0"
notebook = "^7.0.0"
ipykernel = "^6.25.0"
black = "^23.7.0"
isort = "^5.12.0"
click = "^8.1.6"
scipy = { version = "^1.11.1", python = ">=3.11,<3.13" }
xgboost = "^1.7.6"
requests = "^2.31.0"
pyyaml = "^6.0.1"
utils = "^1.0.1"
schema = "^0.7.5"
validators = "^0.21.2"
wikitools = "^1.3"
pyarrow = "^15.0.0"
flake8 = "^7.0.0"
geopandas = "^0.14.3"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
jupyter-black = "^0.3.4"
pyzmq = "^25.1.2"
shap = "^0.45.0"
cashflows = "^0.0.4"
numpy-financial = "^1.0.0"
geopy = "^2.4.1"
newsapi-python = "^0.2.7"

elsapy = "^0.5.1"
scholarly = "^1.7.11"
textblob = "^0.18.0.post0"
asteval = "^1.0.4"
cleanco = "^2.3"
pylint = "^3.3.1"
pyreadr = "^0.5.2"
fuzzywuzzy = "^0.18.0"
levenshtein = "^0.26.1"
plotly-express = "^0.4.1"
geocoder = "^1.38.1"
pyam-iamc = "^3.0.0"
linkchecker = "^10.5.0"
kaleido = "0.2.1"
pyshp = "^2.3.1"
pdfplumber = "^0.11.6"
cpi = "^2.0.6"
beautifulsoup4 = "^4.13.4"
selenium = "^4.33.0"
ipython = "^9.3.0"
shellingham = "^1.5.4"
ipyleaflet = "^0.20.0"
dash = "^3.1.1"
gender-guesser = "^0.4.0"
[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.25.0"
ipython = "^9.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = ".py"
exclude = '''
^/(
  (
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | data
  )/
  | foo.py           # also separately exclude explicitly named files here
)
'''

[tool.isort]
atomic = true
profile = "black"
line_length = 88
py_version = 311
skip_gitignore = true
include_trailing_comma = true
skip_glob = ["tests/data", "profiling"]
extend_skip = [".md", ".json", ".csv"]
known_first_party = ["black", "blib2to3", "blackd", "_black_version"]

[tool.sqlfluff.core]
templater = "jinja"
sql_file_exts = ".sql,.sql.j2,.dml,.ddl"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 95
exclude = ["setup.py", "docs", "build"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 0
quiet = false
whitelist-regex = []
color = true
generate-badge = "."
badge-format = "svg"
name = "projects"
version = "0.1.0"
description = "Data Science Projects"

[tool.pytest.ini_options]
testpaths = ["projects"]
minversion = "6.0"

# pyproject.toml


[tool.poetry.scripts]
cli = "cli.run:cli"
